add cli args update readme
